Use deterministic prefix in loggerName so level can be configured
Without a common hierarchy for the name of a logger, its level cannot be set granularly.  Without this, if I wanted to enable INFO for a JavaScript verticle's logger, I'd have to configure the entire application's root level to INFO or lower.  This still doesn't address the fact that you can't configure a specific logger (because you can't set the logger's name externally), but that would require another deploy-time config param.