Adds two new, public methods: ConnectionBase.getPeerCertificateChain() and HttpServerRequest.getPeerCertificateChain(). The methods return unmodifiable lists of the X509 cert chain for the incoming client request (if the request is ssl).
The cert chain ultimately comes from the netty channel.  This gets passed into ConnectionBase.  The connection is a member of DefaultHttpServerRequest, so we're now able to get the peer certs from the  request via the connection.  This seemed to make more sense than adding the channel dependency to the request class itself.