Adding common hierarchy for JSON classes
JsonArray and JsonObject inherit from JsonElement. JsonElement allows
people to check if it is an Array or Object, convert to those relevant
types. This is also more convenient when dealing with fields that could
be either an Array or an Object; one can simply getElement(), check if
it is an array or object then convert asArray or asObject, instead of
using try-catch.