Disentangle ProxyRepoDownloadTest from RepoDownloadTest
Turns out that the Proxy test only worked because the RepoDownloadTest
had already downloaded the test app and therefore it never actually
hit the proxy server. If it did, it would have noticed that it tried
to hit it with a relative url first (because localhost:9193 is in the
list of repo resolvers) and then tried a wrong url (ending in :9193,
not in :9192).

Run the proxy on another port (9194), fix the request URI using 9193,
change the name of the proxy app to not be the same as the non-proxy
download app.

Both ProxyRepoDownloadTest and RepoDownloadTest are still stateful
tests because if they are run multiple times without cleaning out the
mod-test folder, the test will not hit the actual server the second
and any further time.