During message reply handling, vert.x Event bus bridge decides if the reply contains authentication data (sessionID) which it should cache and then keep using for that socket. The reply is stored as "sessionID" variable in the reply body JSON object.
The code however checks for presence of this property in the message itself (message contains properties such as "type":"send", "address":"xxxxx" etc, not the message body itself. Therefore, such session ID in a reply is never found.

This patch fixes the issue, so that vert.x searches for the session ID in the message body.

Signed-off-by: Michal Boska <michal.boska@erni.sk>