EB send on default context if current context is a worker context
The sending context is not important, because we use the context stored in the handler registration to deliver the message.

So when the sending context is a worker context, fall back to the default context.
Indeed, the "event bus connected" callback (just an example) might not be executed if worker is busy on (user) blocking code.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>