Add support for websocket text framing.
The writeTextMessage method can be used to write a single large text
message as multiple websocket frames.

The binaryMessageHandler or textMessageHandler will receive multi-frame
messages as a single re-assembled buffer or string.

The maximum size of re-assembled messages can be specified with the
maxWebsocketMessageSize http option on both the server and client side.

Messages that are too large to re-assemble are dropped and the
exception handler is invoked with an IllegalStateException.

Issue #1807
Signed-off-by: rmelick <rmelick@cs.hmc.edu>