Fixes intermittent failure in io.vertx.test.core.Http2ClientTest.test100Continue
java.lang.AssertionError: expected:<0> but was:<1>
	at org.junit.Assert.fail(Assert.java:88)
	at org.junit.Assert.failNotEquals(Assert.java:834)
	at org.junit.Assert.assertEquals(Assert.java:645)
	at org.junit.Assert.assertEquals(Assert.java:631)
	at io.vertx.test.core.AsyncTestBase.assertEquals(AsyncTestBase.java:235)
	at io.vertx.test.core.Http2ClientTest.lambda$test100Continue$170(Http2ClientTest.java:1293)

We know the body handler should be called after both the request handler and the continue handler have been called.
But we don't know which one will be called first: in Http2ServerConnection.onHeadersRead, we write continue headers to the response then schedule the request handler execution.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>