Add example to propogate stack traces for EventLoop
Debugging non-blocking/eventloop code can be challenging. Since the
context can move -- the caller loses track of the stack when an
exception is thrown.

This is a working example of necessary code that can be added to the
ContextImpl in order to keep propogating the stack for the caller during
event loop tasks. Similar code can be added easily for the
ExecutorService when executeBlocking is called.

Not familiar enough although perhaps the VertxThread itself should carry
the stack frames?

Signed-off-by: Farid Zakaria <farid.m.zakaria@gmail.com>