make LoggerFactory try its own ClassLoader if TCCCL fails, see #2298
attempts to use LoggerFactory.class.getClassLoader(); as a fallback if
Thread.currentThread().getContextClassLoader() fails to be able to load
the LogDelegateFactory.

This is because the use of getContextClassLoader() can be problematic in
"context ClassLoader sensitive environments", such as OSGi; see e.g.
http://njbartlett.name/2010/08/30/osgi-readiness-loading-classes.html or
http://wiki.eclipse.org/Context_Class_Loader_Enhancements

PS, minor: The renaming of delegateFactory variable to
newDelegateFactory just for clarity and to avoid IDE/javac warning re.
possible confusion with static field of same name.

Signed-off-by: Michael Vorburger <vorburger@redhat.com>