Hydrate vert.x cache for directory on first access
When a file is first access -- when caching is enabled, vert.x will copy
the file to a cache directory.
Subsequent listing of the directory will match the cache however the
contents of the directory will be only the recently accessed file.
To circumvent this bug -- when a file is first accessed, the
FileResolver will cache all other files in the same directory.

Fixes #2126

Signed-off-by: Farid Zakaria <farid.m.zakaria@gmail.com>