Add VertxOption to specify worker pool queue capacity
Add the ability to set the maximum capacity for the worker pool queue at
which point trying to enqueue additional tasks when its full will result
in a RejectedExecutionException

Previously the capacity of the queue was Integer.MAX_VALUE (unbounded)
which can result in a death-spiral for systems (i.e. http services) that
are processing more requests than they can handle.

For instance, as work tasks begin to queue for the workerPool, clients
that have yet to be serviced start retrying their operations. The queue
then continues to grow with clients continuing to retry their operation.
Death spiral.

Signed-off-by: Farid Zakaria <farid.m.zakaria@gmail.com>