Fix FileResolver for OSGi bundle resources
The use of `FileResolver` for a directory without a trailing '`/`' will
fail since `unpackFromBundleURL()` does not handle this case correctly.
By trying `ClassLoader.getResource()` on such a resource path with an
appended '`'`', it can be determined whether or not the resource is
really a directory since a URL will only be returned if the resource
exists and is a directory.

Fixes #2548.

Signed-off-by: Jack O'Sullivan <jackos1998@gmail.com>