Set verticle deployment completion handler early for blocking worker case
Before, the handler for `startFuture` was only registered after `Verticle.start(startFuture)` returned. However in case the verticle is a worker verticle it should be allowed to block indefinitely and still communicate that it has started successfully (before it starts blocking).

There is the slight side-effect of this change that if the verticle start() method throws a *synchronous* exception *after* already having reported successful startup (through `startFuture`), the success will now get propagated to the initiator, when before it wasn't.