The `VertxHandler` does not handle anymore the dispatch to the Vertx Context and instead the connection has full control on the dispatch to the application allowing each protocol to perform application dispatch according to its concurrency.
Support worker contexts in the HTTP server implementation for the HTTP/2 protocol: the implementation uses message passing between the connection event loop and the HTTP streams when the stream are delivered on a worker context, otherwise this will be a simple method call.

Metrics and tracing are now always reported from the IO thread in all cases.

This fixes #3216 .