Explain WriteStream.write internal queue and Handler/Future
The javadoc of write(T, Handler) is unclear whether the Handler
is called after putting T into the internal queue or after
the actual write of T.

Solution:

Copy the javadoc of write(T) to write(T, Handler) to make it clear.

Also mention it in the Future `@return` and Handler `@param` javadoc.

Signed-off-by: Julian Ladisch <eclipse.org-rtn@ladisch.de>