Context.put and .putLocal should remove if value is null
Context.put(key, value) and Context.putLocal(key, value) are implemented
with a ConcurrentHashMap that throws NullPointerException if key or value
is null.

For usability remove the key when putting a null value. This is almost
always what the caller wants.

Add javadoc when put, get, remove, putLocal, getLocal, removeLocal
throw NullPointerException.

Signed-off-by: Julian Ladisch <eclipse.org-rtn@ladisch.de>