New generic KeyStoreOptions (#3592)
Introduce a new generic KeyStoreOptions that can be used when a KeyStore based implementation of JksOptions or TrustOptions. This implementation allows to specify a type that is used when loading
the KeyStore. PfxOptions and TrustOptions have been retrofit to extend this new class and set the type to the appropriate type. The TrustOptions#toJson() method has been removed because it does not seem to belong to TrustOptions and was used mainly for testing purposes. - fixes #3591