Add `ContextManager` service abstraction
The context manager is a new abstraction that allow users to implement
their own service providers to change the default, thread-local-based
mechanism to guarantee that the same context is used from threads that
are not managed by vert.x (for example, when users use the `WebClient`
in an external thread pool).

This new abstraction intends to imitate how the `ContextInternal` API
works.  As the previous code already had a cast from
`ContextInternal.currentContext` to `AbstractContext`, which is an
internal class, the new context manager returns a generic public
instance of `Context` and relies on `VertxImpl` to cast it to whichever
context it wants to use.

The new interface is designed so that implementors don't need to use
internal classes (e.g. `ContextInternal` or `AbstractContext`) to
implement the new interface.

Context manager implementors should not change the stored context
reference and the java doc clearly specifies that. That is to guarantee
context manager implementors don't change the context stored by vert.x
and possibly change its type, causing the internal cast to fail.

Fixes https://github.com/eclipse-vertx/vert.x/issues/3619

Signed-off-by: Jorge Vicente Cantero <jorgevc@fastmail.es>