Allow space in Windows file path
A space is a valid character in Windows file path, but
not at the end of a directory or file name.

The path validation is adjusted to reject a space or dot
at the end of a file or directory name.

The \ character is always replaced by a / character so that
it doesn't get URL encoded.

Signed-off-by: Julian Ladisch <eclipse.org-rtn@ladisch.de>