Vert.x Service Provider bootstrap.
The service provider bootstrap provides a Vert.x SPI mechanism that allows to customize services used by Vert.x and avoid the proliferation of various service loaders. Existing service providers are retrofitted to use the new mechanism (cluster manager, node selector, metrics and tracing).

The current VertxFactory is renamed VertxBuilder and can be used standalone to build a customized Vert.x instance. This builder remains internal as this is not the advocated way to build a Vert.x instance and for most users Vertx.vertx() and Vertx.clusteredVertx() is the supported API.

fixes #3772