VertxThreadFactory service provider (#3800)
The creation of VertxThread instances is delegate to a VertxThreadFactory used by Vert.x when creating a Java ThreadFactory. This factory only creates the thread instance and the internal ThreadFactory will register it against the block thread checker and configure the daemon flag. The VertxThreadFactory extends the VertxServiceProvider so it can be loaded from the classpath using the Java service loader mechanism. The VertxServiceFactory can also be set programmatically on the VertxBuilder. - fixes #3799