Implement an order for JsonFactory in order to define a priority to pick a factory when there are more than one on the classpath. The mechanism is required because the JsonFactory is static and the user cannot specify a factory to use statically.
The JsonFactory method is added an order() method that returns the highest possible integer as default method. Implementations can override this to provide an order. Vert.x does not provide an override mechanism for this but encourages each factory to return a constant that can be overriden by a system property.

Signed-off-by: Julien Viet <julien@julienviet.com>