The current HTTP/1.x client connection ignores any message that is received when there is no request in progress to handle this message. When it happens the connection the connection should be closed.
The Http1xClientConnection now fails the connection when it receives an HTTP message and then is no pending request to handle this message. A couple of tests have been improved because they were triggering this case by immediately sending a response instead of sending a response after the client has sent the request.

Fixes #3831

Signed-off-by: Julien Viet <julien@julienviet.com>