JsonObject/JsonArray provide a copy method that can fail on unknown elements when they cannot be copied by throwing an IllegalStateException. We can provide an overloaded version that uses a Function<Object, ?> to copy unknown elements and avoid failing.
The contract of JsonObject/JsonArray copy is clarified and now says that it will throw IllegalStateException because an element can be safely copied and this element should not be ignored by the copy method.

An overloaded version of copy that takes a clone function is added and this function is called for unknown elements, providing the user the opportunity to copy such JsonObject/JsonArray easily.

The actual copy method delegates to the new version with a clone function that simply throws IllegalStateException.

fixes #3852
fixes #3823

Signed-off-by: Julien Viet <julien@julienviet.com>