The reply handler result promise can be completed more than one time resulting in internal failures, this promise should be completed only once. The reply handler delivery performs a lookup on the reply handler and then schedule the message delivery on its event loop. There is a race happening when the reply timeout fires after the lookup and before the message delivery.
We can use the reply handler timeout execution/cancellation to guard against multiple completion of this promise. On a reply, the result promise is completed when the reply timer can be cancelled. On a timeout the result promise is also completed.

In addition the internal reply handler flow is a bit improved and the ReplyHandler now implements Handler<Long> to avoid creating a lambda. The reply handler was unregistered twice (with no side effects) on a reply failure, now it is only once.

fixes #3862
fixes #3859

Signed-off-by: Julien Viet <julien@julienviet.com>