An h2c 'rejected' upgrade (e.g an HTTP/1 server totally ignoring h2c upgrade) can remain in the pool when the connection is closed.
This is a regression introduced when the HTTP client was decoupled from the pool implementation.

The upgraded h2c client connection assumes the upgrade always succeeds and instead when the upgrade does not happen should set an eviction handler to properly evict the connection when the connection is closed.

fixes #1946