Do not deliver remote message to localOnly consumers
Fixes #3959

Consider a node that has both local and clustered consumers registered for an address.
When a remote message is received (sent from another node), calling deliverMessageLocally would lead to message being possibly delivered to one of the local consumers.
This is not what users expect.

With a variation of ConcurrentCyclicSequence, we can make sure we deliver:

- to all consumers when a message was sent from this node
- to clustered consumers only when a message was sent from another node

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>