Add aliasPassword property to keystore options
Fixes #3970

In #3933 we added an alias property to keystore options that allows to choose a keystore entry when there are several instead of letting the JVM selecting the first one.

This is a follow-up change that allows to specify a password for the alias when it is different than the store's password.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>