Context: allow storage and retrieval of instances of any type
Closes #3952

The Context interface used to restrict storage and retrieval of objects in context to String instances.

But the underlying storage permits to use instances of any type.

With this change, users can now store any type of objects in the context, without having to cast to ContextInternal and work with data maps directly.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>