Don't create a Promise from the finalizer
This can result in spawning a new IO thread, which will then inherit the
TCCL and ThreadGroup of the finalizer thread.

This can still be an issue if the client is not closed properly, but
currently there is no way to avoid it even if you do close it properly.

HttpClientImpl does not need a finalizer, as all it does in invoke the
NetClientImpl close, and NetClientImpl has its own finalizer.

Signed-off-by: Stuart Douglas <sdouglas@redhat.com>