HTTP server request events wrong metrics events for pipelined requests (#4091)
HTTP server request events will notify observability events incorrectly when processing pipelined requests. The request begin event will be omitted and the request end event will be emitted independently of the application request handling.

The events should be emitted when the application process the request and not when the server receives the event.

see #4090