The Vert.x HTTP server wraps the ChannelHandlerContext of the HttpObjectEncoder to force the encoder to allocate direct buffers for the response. Which is only useful for clear text servers.
We remove this wrapping and instead a clear text server will prefer to use direct buffers instead of the partial allocator.

When using clear text the server will continue allocating direct buffers, there is no consequence for the decoder since the buffers are allocated by the event loop and are direct and they need to be copied anyway by the server when wrapping the ByteBuf with a Vert.x Buffer.

With a TLS channel, the encoder now allocates on the heap but such buffers are anyway encoded by the SSL handler.