Make `BufferImpl` respect the `Buffer` interface
If a `Buffer` is passed to `BufferImpl`, `BufferImpl` should respect that the passed `Buffer` instance might not be of type `BufferImpl` but of another class, that implemements the `Buffer` interface. This is achieved by checking the type and either using the `byteBuf()` method to get the internal byte buffer of `BufferImpl` (as if a object is passed via a parameter, it is considered a code smell to access its private fields, especially in case a method is available) or by calling `getByteBuf()` in case the passed object implements the `Buffer` interface.