Implement message codec lookup that performs super class lookup.
When a codec is looked up, it will lookup the super class when the lookup returns null.

- Object.class is not considered for lookup
- only super class is attempted, not implemented interfaces
- raw class lookup are supported, no support for parameterized types
- system codec are resolved before other codecs

The implementation also simplifies the internals of the codec manager and the `MessageCodec` method `systemCodecID` is now a default methods that returns -1.

Registering a default codec for java.lang.Object will throw IllegalArgumentException

fixes #4120