Change how NetSocketInternal delivers messages to the message handler. The current implementation delivers the message through the pending inbound buffer that will then dispatch the message to the message handler or the data handler (through a custom data message handler), buffers are copied and released before the message is written in the inbound buffer to release direct memory when using in regular socket mode.
The new implementation moves the inbound buffer use to the data message handler along with the buffer copy and release. Any other message handler therefore handles the message directly without the inbound buffer and therefore NetSocket back-pressure does not apply anymore, in addition plain ByteBuf needs now to be released by the message handler, giving the opportunity to the handler to avoid the message copy when that is not needed.