Serializer may throw ConcurrentModificationException when used from worker verticle
Fixes #4128

When invoked from a worker verticle, Serializer queues messages on the event loop of the worker context.
But when the serialized task completes, it executes on the worker context, ie the worker thread.
Therefore, the queues map can be modified concurrently.

With this change, we make sure the Serializer code is only invoked from an event loop.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>