Several fixes necessary to implement WebSocket upgrade support in a reverse proxy implemented with Vert.x
When using the HTTP client connect API, the HTTP/1.1 implementation incorrectly case sensitively compares the connection header to the Upgrade value and therefore does not always remove the HTTP handlers on the pipeline to handle correctly the ByteBuf to the upgraded NetSocket.

The HTTP/1.1 server will only provide a NetSocket when the HTTP CONNECT method is used, it should also work when the client is requesting an HTTP upgrade. Therefore the HTTP server will return a valid NetSocket if and only if the request is a CONNECT or an HTTP connection upgrade, depending on the case a 200 or 101 response will be sent.

fixes #4156
fixes #4152