Do not close connection on http client request connect (#4161)
The HttpClientRequest ends the stream when the request has been sent with the connect API, which close the underlying TCP socket. Instead the request should not end the stream as this will be done by the NetSocket after the upgrade. This change is necessary to support WebSocket https://datatracker.ietf.org/doc/html/draft-ietf-hybi-thewebsocketprotocol-00 in a reverse proxy as this version of WebSocket has a body.