HTTP client will not update correctly the underlying TCP client metrics
The HTTP client will create a TCPMetrics and a HttpClientMetrics and only update the HttpClientMetrics, leaving an incorrect TCPMetrics.

Instead the HTTP client should create a NetClient passing its own HttpClientMetrics as argument to have a single metrics instance and let the NetClient signal the connection event. When the NetSocket is promotted to an HttpClientConnection, the created metrics will be reused instead of being created again.

fixe #4200