Pool can deliver lease with null connection.
A race can happen in the pool when it delivers a lease. Leases are created as pool post actions and sometimes can read a connection from a slot that became null in the meantime the post action is executed.

Leases must be created before in pool tasks so they read the connection and then they are delivered with a non null (but potentially closed) connection.

fixes #4221