The implementation of the default invalid request handler only send a 400 response (bad_request) in some situation otherwise it will simply close the connection without sending a response message.
It should instead always send a response message (400 when nothing more specific can be sent) and then close the connection since the Netty decoder is in an un-recoverable state.

This commit changes the default invalid request handler to senda 400 response when no specific status code can be sent and then close the response.