Return failed future when executeBlocking task is rejected
Fixes #4209

When Vert.x is closed, there's a small timeframe which allows tasks to be submitted on worker pool but they will never get executed because the pool is closed.

Instead of throwing the runtime exception, we can return a failed future and give the user a chance to deal with it.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>