Local consumer should never get a message sent or published remotely
Fixes #4116

When a node has a mix of local and clustered consumers registered to a single address, a message sent from a remote node may be delivered to the local consumers, which is unexpected.

This happens because when the node receives the clustered message, it invokes deliverMessageLocally which does not inspect the type of handlers.

So, in case of point to point messaging, we need to iterate over the content of the sequence of handlers until we find one which is either a reply handler or not local only.
For pub/sub, we need to filter out local-only consumers.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>