Let users define a codec to use when no codec was found
Register codecs as usual
Provide a function to the EventBus
The function is invoked when no codec has been found for message body and must return a user codec name.

Serializable and ClusterSerializable codecs are used as a last resort so that users can decide to encode objects implementing these interfaces as they wish.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>