HTPP client ignores direct H2C preface errors
The HttpClient assumes that after a directH2C connection (without HTTP upgrade) has been established with the server, it will always be successfully interacting with an H2 server.

When a server sends invalid data (from the client perspective), e.g a plain HTTP server, the client should handle this case and fail the connection result.

A connect future is added to VertxHttp2ConnectionHandler that is completed when the preface has been read and failed when a connection error happens before the settings have been read. When the client receives a corrupted settings frame, the connection error is triggered and the future is failed.

The connector now interacts with a promise of connection instead of the previous handler, removing the assumption the connection will always be established.

see #4367 #4254