The thread factory SPI should be able to use the thread factory wrapper when it is created.
The wrapper of the vertx factory that is passed to the SPI relies on the vertx instance and it is not properly constructed when the thread factory SPI uses the wrapper when it is initialized.

The wrapper is now initialized with the resources it needs instead of retrieving them from the vertx instance.

fixes #4471