The JsonParser class does not progress as much as it could when the processing faulty input.
When the JsonParser process malformed input, it does halt the parsing and handle the error leaving unprocessed tokens in the underlying Jackson parser. These tokens will be processed when the JsonParser handle another input. When the Jackson parser fails we handle the error and then give up and reported behavior is a consequence.

The JsonParser has been modified to continue when a malformed input is reported instead of stopping after the error has been processed. Errors are collected and delivered after we all tokens have been processed.

fixes #4338