The pool asynchronously removes connection slots when they are identified as removable. This create races, e.g a connection can be evicted from the pool and then acquired before the removal happens.
Connection slots are now removed synchronously when the pool identifies connections to be removed immediately within the pool logic. The post action removal are now enqueued in the main post action (when there is one), the combiner executor has been modified to take in account the fact that a post action might next actions to execute.