Implement an MPSC queue named OutboundWriteQueue which provides a concurrent multi producers back-pressured queue fronting a single back-pressured consumer.
Producers emission flow cannot reliably be controlled by the consumer back-pressure probe since producers emissions are transferred to the consumer thread: the consumer back-pressure controller does not take in account the inflight elements between producers and consumer. This queue is designed to provide reliable signals to control producers emission based on the consumer back-pressure probe and the number of inflight elements between producers and the consumer.